fix(macos): third wave of macOS compile fixes
- Add MEMORYSTATUS/GlobalMemoryStatus, LARGE_INTEGER/QueryPerformanceCounter/
  QueryPerformanceFrequency, Sleep, GetDoubleClickTime, VK_RETURN, DeleteFile,
  GetLastError, FormatMessage(W), _spawnl/_P_NOWAIT, OSVERSIONINFO/GetVersionEx,
  LOCALE_SYSTEM_DEFAULT/GetDateFormat, SHGetSpecialFolderLocation/SHGetPathFromIDList
  stubs to compat.h
- Add itoa, _wtoi stubs to string_compat.h / wchar_compat.h
- Fix ~30 pointer-to-smaller-type cast errors across GUI/logic files by
  inserting (uintptr_t) intermediate cast
- Gate _fpreset/_controlfp FP mode setup behind #ifdef _WIN32 in GameLogic.cpp
- Gate WIN32_FIND_DATA user-data path block behind #ifdef _WIN32 in Image.cpp
- Rename 'pause' variable to 's_pause' in InGamePopupMessage.cpp to avoid
  conflict with POSIX pause() function
- WindowMsgData typedef changed to uintptr_t in GameWindow.h for 64-bit safety

+++ Core/GameEngine/Source/Common/INI/INI.cpp
@@ -669,7 +669,7 @@ void INI::parseBool( INI* ini, void * /*instance*/, void *store, const void* /*u
 void INI::parseBitInInt32( INI *ini, void *instance, void *store, const void* userData )
 {
 	UnsignedInt* s = (UnsignedInt*)store;
-	UnsignedInt mask = (UnsignedInt)userData;
+	UnsignedInt mask = (UnsignedInt)(uintptr_t)userData;
 
 	if (INI::scanBool(ini->getNextToken()))
 		*s |= mask;
